feat: update aws-serverless-experss to @vendia/serverless-express and switching lambda support from fastify to express
resolves https://github.com/frouriojs/create-frourio-app/issues/230
