+++ server/pretense/package.json
@@ -37,7 +37,6 @@
     "@testing-library/react": "^12.1.2",
     "@tsconfig/svelte": "^3.0.0",
     "@types/aws-lambda": "^8.10.92",
-    "@types/aws-serverless-express": "^3.3.5",
     "@types/busboy": "^1.3.0",
     "@types/compression": "^1.7.2",
     "@types/cors": "^2.8.12",
@@ -50,9 +49,9 @@
     "@types/react": "^17.0.39",
     "@typescript-eslint/eslint-plugin": "^5.10.2",
     "@typescript-eslint/parser": "^5.10.2",
+    "@vendia/serverless-express": "^4.5.3",
     "@vue/test-utils": "^1.3.0",
     "aspida": "^1.7.1",
-    "aws-serverless-express": "^3.4.0",
     "axios": "^0.25.0",
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "^10.1.0",
