+++ server/templates/server/@server=express/entrypoints/@orm=prisma/@serverless=lambda/@server-dep
@@ -1,0 +1,1 @@
+@vendia/serverless-express
