+++ server/templates/server/@server=express/entrypoints/@orm=prisma/@serverless=lambda/@server-dev-dep
@@ -1,0 +1,1 @@
+@types/aws-lambda
