+++ server/templates/server/@server=express/entrypoints/@orm=prisma/@serverless=lambda/lambda.ts
@@ -1,0 +1,5 @@
+import { init } from "$/service/app";
+import serverlessExpress from "@vendia/serverless-express";
+
+const app = init();
+export const handler = serverlessExpress({ app });
