+++ server/templates/server/@server=express/entrypoints/@orm=prisma/@serverless=lambda/lambda_migration.ts
