--- server/templates/server/@server=fastify/entrypoints/@orm=prisma/@serverless=lambda/@server-dep
@@ -1,1 +1,0 @@
-aws-serverless-express
