--- server/templates/server/@server=fastify/entrypoints/@orm=prisma/@serverless=lambda/@server-dev-dep
@@ -1,2 +1,0 @@
-@types/aws-lambda
-@types/aws-serverless-express
